CCU Heartbeat and HES (Meter Client) Rewrite
