+++ README.md
@@ -31,8 +31,6 @@ service supervisor restart
 > git clone https://github.com/manqala/hearbeat_server.git
 (env)> pip install -e heartbeat_server
 ```
-Ensure that the current user has write permissions in the folder where supervisor is installed (/etc/supervisor/conf.d/)
-
 4. Setup app configuration and edit 'config.json' as required
 ```
 (env)> python -m heartbeat_server --setup
