+++ heartbeat_server/logger.py
@@ -3,18 +3,19 @@ import pathlib
 import sys
 from logging.handlers import RotatingFileHandler
 
-
-LOG_LEVEL = logging.DEBUG
+DEFAULT_LOG_LEVEL = logging.DEBUG
 LOG_FILENAME = './logs/app.log'
 
 
-def get_logger():
+def get_logger(log: dict = None):
+    log_file = log.get('file') \
+        if log.get('file') \
+        else DEFAULT_LOG_LEVEL
     logger = logging.getLogger('heartbeat_server')
     if not len(logger.handlers):
         formatter = logging.Formatter('[%(levelname)s] %(asctime)s | %(name)s:\n%(message)s\n')
-
         try:
-            folder = pathlib.Path(LOG_FILENAME).parent
+            folder = pathlib.Path(log_file).parent
             folder.mkdir(parents=True, exist_ok=True)
         except FileExistsError:
             pass
@@ -25,7 +26,7 @@ def get_logger():
         console_handler = logging.StreamHandler(sys.stdout)
         console_handler.setFormatter(formatter)
 
-        logger.setLevel(LOG_LEVEL)
+        logger.setLevel(log.get('level'))
         logger.addHandler(file_handler)
         logger.addHandler(console_handler)
         logger.propagate = False
